+++ linkerd/core/src/main/scala/io/buoyant/linkerd/Announcer.scala
@@ -1,0 +1,11 @@
+package io.buoyant.linkerd
+
+import com.twitter.finagle.{Announcer => FAnnouncer, Path}
+
+abstract class Announcer extends FAnnouncer {
+  def concreteName(name: String): Path
+}
+
+
+
+